ref(utils/math): Changed parameter type for 16-bit rotation functions

+++ src-server/shared/utils/math/main.h
@@ -27,12 +27,12 @@ static inline uint8_t rotr8(const uint8_t x, uint8_t n) {
     return (x >> (n & 7)) | (x << ((8 - n) & 7));
 }
 
-static inline uint16_t rotl16(const uint16_t x, uint8_t n) {
+static inline uint16_t rotl16(const uint16_t x, uint16_t n) {
     n &= 0x10;
     return (x << (n & 15)) | (x >> ((16 - n) & 15));
 }
 
-static inline uint16_t rotr16(const uint16_t x, uint8_t n) {
+static inline uint16_t rotr16(const uint16_t x, uint16_t n) {
     n &= 0x10;
     return (x >> (n & 15)) | (x << ((16 - n) & 15));
 }
